+++ source/mysql/result.d
@@ -15,6 +15,7 @@ import std.variant;
 import mysql.common;
 import mysql.connection;
 import mysql.protocol.commands;
+import mysql.protocol.extra_types;
 import mysql.protocol.packets;
 
 /++